Remove old place legaliser, set placement constraints instead (currently ignored by placer)
